Move utility function to utils file
